# services/ucs/openapi.yaml
components:
  schemas:
    BoxCreationRequest:
      description: Request body for creating a new FileUploadBox.
      properties:
        storage_alias:
          description: The storage alias to use for this upload box
          title: Storage Alias
          type: string
      required:
      - storage_alias
      title: Box Creation Request
      type: object
    BoxUpdateRequest:
      description: Request body for updating a FileUploadBox.
      properties:
        state:
          description: Updated state
          enum:
          - open
          - locked
          - archived
          title: State
          type: string
        version:
          description: The expected current version of the box (for optimistic locking)
          title: Version
          type: integer
      required:
      - state
      - version
      title: Box Update Request
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    HttpBoxNotFoundError:
      additionalProperties: false
      properties:
        data:
          $ref: '#/components/schemas/HttpBoxNotFoundErrorData'
        description:
          description: A human readable message to the client explaining the cause
            of the exception.
          title: Description
          type: string
        exception_id:
          const: boxNotFound
          title: Exception Id
          type: string
      required:
      - data
      - description
      - exception_id
      title: HttpBoxNotFoundError
      type: object
...
    HttpBoxVersionError:
      additionalProperties: false
      properties:
        data:
          $ref: '#/components/schemas/HttpBoxVersionErrorData'
        description:
          description: A human readable message to the client explaining the cause
            of the exception.
          title: Description
          type: string
        exception_id:
          const: boxVersionOutdated
          title: Exception Id
          type: string
      required:
      - data
      - description
      - exception_id
      title: HttpBoxVersionError
      type: object
...
    HttpOrphanedMultipartUploadError:
      additionalProperties: false
      properties:
        data:
          $ref: '#/components/schemas/HttpOrphanedMultipartUploadErrorData'
        description:
          description: A human readable message to the client explaining the cause
            of the exception.
          title: Description
          type: string
        exception_id:
          const: orphanedMultipartUpload
          title: Exception Id
          type: string
      required:
      - data
      - description
      - exception_id
      title: HttpOrphanedMultipartUploadError
      type: object
...
    HttpUnknownStorageAliasError:
      additionalProperties: false
      properties:
        data:
          $ref: '#/components/schemas/HttpUnknownStorageAliasErrorData'
        description:
          description: A human readable message to the client explaining the cause
            of the exception.
          title: Description
          type: string
        exception_id:
          const: noSuchStorage
          title: Exception Id
          type: string
      required:
      - data
      - description
      - exception_id
      title: HttpUnknownStorageAliasError
      type: object
    HttpUnknownStorageAliasErrorData:
      additionalProperties: true
      properties: {}
      title: HttpUnknownStorageAliasErrorData
      type: object
...
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http
info:
  description: A service managing uploads of file objects to an S3-compatible Object
    Storage.
  title: Upload Controller Service
  version: 11.0.0
openapi: 3.1.0
paths:
  /boxes:
    post:
      description: 'Create a new FileUploadBox.


        Requires CreateFileBoxWorkOrder token and only allowed for Data Stewards via
        the UOS.

        Request body should contain the storage alias to use for uploads within the
        box.

        Returns the box_id of the newly created FileUploadBox.'
      operationId: createBox
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoxCreationRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                format: uuid4
                title: Response Createbox
                type: string
          description: The box_id of the newly created FileUploadBox
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpUnknownStorageAliasError'
          description: 'Exceptions by ID:

            - noSuchStorage: The storage node for the given alias does not exist.'
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Create a new FileUploadBox
      tags:
      - UploadControllerService
  /boxes/{box_id}:
    patch:
      description: 'Update a FileUploadBox to lock it, unlock it, or archive it.


        Request body must contain a `state` field indicating the target state of the
        box,

        as well as a `version` field indicating the expected current version of the
        box.

        Requires ChangeFileBoxWorkOrder token from the UOS. Users are only allowed
        to lock

        the box; a Data Steward role is required to do everything else.'
      operationId: updateBox
      parameters:
      - in: path
        name: box_id
        required: true
        schema:
          format: uuid4
          title: Box Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoxUpdateRequest'
        required: true
      responses:
        '204':
          description: FileUploadBox successfully updated
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpBoxNotFoundError'
          description: 'Exceptions by ID:

            - boxNotFound: The FileUploadBox with the given ID does not exist.'
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpBoxVersionError'
          description: 'Exceptions by ID:

            - boxVersionOutdated: The requested FileUploadBox version is outdated.'
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Update a FileUploadBox (lock/unlock/archive)
      tags:
      - UploadControllerService
  /boxes/{box_id}/uploads:
    get:
      description: 'Retrieve list of FileUploads for a FileUploadBox.


        Returns the list of FileUploads for completed uploads in the specified box.

        Requires ViewFileBoxWorkOrder token from the UOS.'
      operationId: getBoxUploads
      parameters:
      - in: path
        name: box_id
        required: true
        schema:
          format: uuid4
          title: Box Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/FileUpload'
                title: Response Getboxuploads
                type: array
          description: List of file IDs for completed uploads in the box
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpBoxNotFoundError'
          description: 'Exceptions by ID:

            - boxNotFound: The FileUploadBox with the given ID does not exist.'
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Retrieve list of file IDs for box
      tags:
      - UploadControllerService
    post:
      description: 'Add a new FileUpload to an existing FileUploadBox.


        Creates a new file upload within the specified box with the provided alias,
        checksum, and size.

        Initiates a multipart upload and returns the file ID for the newly created
        upload.

        Requires a CreateFileWorkOrder token from the WPS.'
      operationId: createFileUpload
      parameters:
      - in: path
        name: box_id
        required: true
        schema:
          format: uuid4
          title: Box Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileUploadCreationRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                format: uuid4
                title: Response Createfileupload
                type: string
          description: The file_id of the newly created FileUpload
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpUnknownStorageAliasError'
          description: 'Exceptions by ID:

            - noSuchStorage: The storage node for the given alias does not exist.'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpBoxNotFoundError'
          description: 'Exceptions by ID:

            - boxNotFound: The FileUploadBox with the given ID does not exist.'
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpOrphanedMultipartUploadError'
          description: 'Exceptions by ID:

            - orphanedMultipartUpload: A multipart upload is already in progress for
            this file but cannot be aborted. Request file deletion and then attempt
            the upload again.'
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Add a new FileUpload to an existing FileUploadBox
      tags:
      - UploadControllerService
  /boxes/{box_id}/uploads/{file_id}:
    delete:
      description: 'Remove a FileUpload from the FileUploadBox.


        Deletes the FileUpload and tells S3 to cancel the multipart upload if applicable.

        Requires a DeleteFileWorkOrder token from the WPS.'
      operationId: removeFileUpload
      parameters:
      - in: path
        name: box_id
        required: true
        schema:
          format: uuid4
          title: Box Id
          type: string
      - in: path
        name: file_id
        required: true
        schema:
          format: uuid4
          title: File Id
          type: string
      responses:
        '204':
          description: FileUpload removed successfully
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpUnknownStorageAliasError'
          description: 'Exceptions by ID:

            - noSuchStorage: The storage node for the given alias does not exist.'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpS3UploadDetailsNotFoundError'
          description: 'Exceptions by ID:

            - s3UploadDetailsNotFound: S3 upload details for the file could not be
            found.'
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpBoxStateError'
          description: 'Exceptions by ID:

            - boxStateError: The FileUploadBox''s state precludes the requested action.'
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpUploadAbortError'
          description: 'Exceptions by ID:

            - uploadAbortError: There was an error aborting the s3 multipart upload.'
      security:
      - HTTPBearer: []
      summary: Remove a FileUpload from the FileUploadBox
      tags:
      - UploadControllerService
    patch:
      description: 'Complete file upload by instructing S3 to finalize the multipart
        upload.


        Concludes the file upload process in UCS by instructing S3 to complete the

        multipart upload for the specified file.

        Requires a CloseFileWorkOrder token from the WPS.'
      operationId: completeFileUpload
      parameters:
      - in: path
        name: box_id
        required: true
        schema:
          format: uuid4
          title: Box Id
          type: string
      - in: path
        name: file_id
        required: true
        schema:
          format: uuid4
          title: File Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileUploadCompletionRequest'
        required: true
      responses:
        '204':
          description: File upload completed successfully
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpChecksumMismatchError'
          description: 'Exceptions by ID:

            - checksumMismatch: The user-supplied encrypted checksum doesn''t match
            S3.'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpFileUploadNotFoundError'
          description: 'Exceptions by ID:

            - fileUploadNotFound: The FileUpload could not be found.'
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpBoxStateError'
          description: 'Exceptions by ID:

            - boxStateError: The FileUploadBox''s state precludes the requested action.'
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpUploadCompletionError'
          description: 'Exceptions by ID:

            - s3UploadCompletionFailure: There was an error completing the s3 multipart
            upload. Delete the file from the file upload box and retry.'
      security:
      - HTTPBearer: []
      summary: Complete file upload
      tags:
      - UploadControllerService
...